requirements.txt
asgiref==3.6.0
Django==4.1.5
django-dotenv==1.4.2
djangorestframework==3.14.0
pytz==2022.7
sqlparse==0.4.3
